Doc: ship cmake example with 'make distcheck'
The file Makefile.am in the examples directory was modified. The
modifications were done to include the new cmake example. The new cmake
example is built when make is invoke in the root directory. Further,
this new example also ships in the tarball when the latter is
created ("make distcheck").

Link: https://lists.lttng.org/pipermail/lttng-dev/2016-June/026232.html
Signed-off-by: Sebastien Boisvert <sboisvert@gydle.com>
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>