Fix: use header file in lttng-ust source
When building the CMake example, the lttng-ust header file and the
lttng-ust shared library must be available, even if lttng-ust is not
installed on the system.

This change was tested on Ubuntu 14.04.

Everything related to lttng was uninstalled, in order to assess whether
the files from the source distribution could be found.

    $ dpkg -l|grep lttng|grep ii|wc -l
    0

The build succeeds, although LTTngUST is not found by FindLTTngUST.cmake.

CMake reports that:

    -- Could NOT find LTTngUST (missing:  LTTNGUST_LIBRARIES LTTNGUST_INCLUDE_DIRS)

One issue is that FindLTTngUST.cmake does not look for the lttng-ust
shared library in the lttng-ust source distribution.

Link: https://bugs.lttng.org/issues/1042
Cc: Ricardo Nabinger Sanchez <rnsanchez@gmail.com>
Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Cc: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Sebastien Boisvert <sboisvert@gydle.com>